Chore: move field to fields module
