+++ public/app/features/logs/components/otel/formats.ts
@@ -1,6 +1,6 @@
 import { LogRowModel } from '@grafana/data';
 
-import { LOG_LINE_BODY_FIELD_NAME } from '../fieldSelector/logFields';
+import { LOG_LINE_BODY_FIELD_NAME, OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from '../fieldSelector/logFields';
 import { LogListModel, NEWLINES_REGEX } from '../panel/processing';
 
 /**
@@ -103,8 +103,6 @@ const OTEL_RESOURCE_ATTRS_REGEX =
   /^(aws_|cloud_|cloudfoundry_|container_|deployment_|faas_|gcp_|host_|k8s_|os_|process_|service_|telemetry_|cluster$|namespace$|pod$)/;
 const OTEL_LOG_FIELDS_REGEX =
   /^(flags|observed_timestamp|severity_number|severity_text|span_id|trace_id|detected_level)$/;
-
-export const OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME = '___OTEL_LOG_ATTRIBUTES___';
 
 export function getOtelAttributesField(log: LogListModel, wrapLogMessage: boolean) {
   const additionalFields = Object.keys(log.labels).filter(
