+++ public/app/features/logs/components/panel/LogLine.tsx
@@ -20,8 +20,7 @@ import { Button, Icon, Tooltip } from '@grafana/ui';
 
 import { LogLabels } from '../LogLabels';
 import { LogMessageAnsi } from '../LogMessageAnsi';
-import { LOG_LINE_BODY_FIELD_NAME } from '../fieldSelector/logFields';
-import { OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from '../otel/formats';
+import { LOG_LINE_BODY_FIELD_NAME, OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from '../fieldSelector/logFields';
 
 import { HighlightedLogRenderer } from './HighlightedLogRenderer';
 import { useLogDetailsContext } from './LogDetailsContext';
